Move DefaultAutoNetTransport to AutoNetServerImpl
