+++ src/autonet/AutoNetServer.cpp
@@ -2,33 +2,6 @@
 #include "stdafx.h"
 #include "AutoNetServer.h"
 
-DefaultAutoNetTransport::DefaultAutoNetTransport(void){}
-DefaultAutoNetTransport::~DefaultAutoNetTransport(void){}
-
-void DefaultAutoNetTransport::Start(void) {
-  
-}
-
-void DefaultAutoNetTransport::Stop(void) {
-  
-}
-
-void DefaultAutoNetTransport::Send(connection_hdl hdl, const std::string& msg){
-  
-}
-
-void DefaultAutoNetTransport::OnOpen(std::function<void(connection_hdl)> fn) {
-  
-}
-
-void DefaultAutoNetTransport::OnClose(std::function<void(connection_hdl)> fn) {
-  
-}
-
-void DefaultAutoNetTransport::OnMessage(std::function<void(connection_hdl, std::string)> fn) {
-  
-}
-
 AutoNetServer::AutoNetServer():
   CoreThread("AutoNetServer")
 {}
